build(warnings): enable -Wall -Wextra -Wpedantic on all library targets and fix all warnings
Link viper_common_opts to all 18 C++ library targets and add C warning
flags to the runtime object library loop. Fix every warning surfaced:

- Remove 15 unused functions (dead Huffman encoder, GC helpers, etc.)
- Remove 11 unused variables and dead code blocks
- Rename 15 shadowing variables in frontends, codegen, and runtime
- Fix 21 cast-qual warnings via (uintptr_t) intermediate casts
- Remove 8 unreachable code blocks after [[noreturn]] traps
- Add explicit {} initializers for 5 missing struct fields
- Restructure SNPRINTF_OR_FAIL macro to avoid GNU ##__VA_ARGS__
- Add memcpy-based fnptr_to_voidptr() for POSIX thread APIs
- Suppress format-nonliteral in 3 intentional variadic wrappers
- Initialize 2 sometimes-uninitialized variables
- Fix 1 signed/unsigned comparison
- Globally suppress -Wno-gnu-label-as-value (VM computed goto)
  and -Wno-overloaded-virtual (Apple Clang visitor pattern)

Build: zero warnings with -Werror, 1209/1209 tests pass.

+++ src/runtime/arrays/rt_array.c
@@ -50,7 +50,7 @@ extern _Noreturn void rt_trap(const char *msg);
 /// @return Header describing the allocation, or `NULL` when @p payload is null.
 rt_heap_hdr_t *rt_arr_i32_hdr(const int32_t *payload)
 {
-    return payload ? rt_heap_hdr((void *)payload) : NULL;
+    return payload ? rt_heap_hdr((void *)(uintptr_t)payload) : NULL;
 }
 
 /// @brief Trap execution due to an out-of-bounds access.
@@ -61,8 +61,7 @@ rt_heap_hdr_t *rt_arr_i32_hdr(const int32_t *payload)
 void rt_arr_oob_panic(size_t idx, size_t len)
 {
     char buf[128];
-    snprintf(buf, sizeof(buf),
-             "rt_arr_i32: index %zu out of bounds (len=%zu)", idx, len);
+    snprintf(buf, sizeof(buf), "rt_arr_i32: index %zu out of bounds (len=%zu)", idx, len);
     rt_trap(buf);
 }
