+++ src/runtime/oop/rt_option.c
@@ -457,5 +457,5 @@ rt_string rt_option_to_string(void *obj)
             break;
     }
 
-    return rt_string_from_bytes(buf, (int64_t)strlen(buf));
+    return rt_string_from_bytes(buf, strlen(buf));
 }
